Elixir v1.3.4 Behaviour

This module has been deprecated.

Instead of defcallback, one can simply use @callback. Instead of defmacrocallback, one can simply use @macrocallback. Instead of __behaviour__(:callbacks), one can simply use behaviour_info(:callbacks).

Macros

defcallback(spec)

Defines a function callback according to the given type specification.

defmacrocallback(spec)

Defines a macro callback according to the given type specification.