ExUnit v1.3.4 ExUnit.TestCase

A struct that keeps information about the test case.

It is received by formatters and contains the following fields:

:name - the test case name

:state - the test error state (see ExUnit.state)

:tests - all tests for this case

Types

t()

t :: %ExUnit.TestCase{name: module, state: ExUnit.state, tests: [ExUnit.Test.t]}