Elixir v1.1.1 File.Stream

Defines a File.Stream struct returned by File.stream!/3.

The following fields are public:

path - the file path

modes - the file modes

raw - a boolean indicating if bin functions should be used

line_or_bytes - if reading should read lines or a given amount of bytes

Types

t :: %File.Stream{line_or_bytes: term, modes: term, path: term, raw: term}