Elixir v1.0.5 IO.Stream

Defines a IO.Stream struct returned by IO.stream/2 and IO.binstream/2.

The following fields are public:

device - the IO device

raw - a boolean indicating if bin functions should be used

line_or_bytes - if reading should read lines or a given amount of bytes

Types

t :: %IO.Stream{device: term, line_or_bytes: term, raw: term}