Mix v1.3.0 Mix.Local.Installer behaviour

This module implements pieces of functionality shared by the archive- and escript-related tasks.

Functions

install(arg, argv, switches)

Specs

install({module, atom}, OptionParser.argv, Keyword.t) :: boolean

Common implementation of installation for archives and escripts.

Relies on a few callbacks provided by respective callback modules for customizing certain steps in the installation process.

print_list(type, items)

Specs

print_list(atom, [String.t]) :: :ok

Prints a list of items in a uniform way. Used for printing the list of installed archives and escripts.

uninstall(type, argv)

Specs

uninstall(atom, OptionParser.argv) :: boolean

A common implementation for uninstalling archives and scripts.

Callbacks

check_path_or_url(arg0)

Specs

check_path_or_url(String.t) :: :ok | {:error, String.t}

Checks that the argument given to install is supported by the respective module.

find_previous_versions(arg0, arg1)

Specs

find_previous_versions(String.t, Path.t) :: [Path.t]

Returns a list of already installed version of the same archive or escript.

install(dest, contents, previous)

Specs

install(dest :: Path.t, contents :: binary, previous :: [Path.t]) :: :ok

The installation itself.