Mix v1.3.4 Mix.Tasks.App.Start

Starts all registered apps.

The application is started by default as temporary. In case :start_permanent is set to true in your project configuration or the --permanent flag is given, it is started as permanent, which guarantee the node will shutdown in case the application crashes permanently.

Configuration

:start_permanent - the application and all of its children applications are started in permanent mode

:consolidate_protocols - when true, loads consolidated protocols before start. The default value is true.

:elixir - matches the current elixir version against the given requirement

Command line options

--force - force compilation regardless of compilation times

--temporary - start the application as temporary

--permanent - start the application as permanent

--no-compile - do not compile even if files require compilation

--no-protocols - do not load consolidated protocols

--no-archives-check - do not check archives

--no-deps-check - do not check dependencies

--no-elixir-version-check - do not check Elixir version

--no-start - do not start applications after compilation

Functions

run(args)

run(OptionParser.argv) :: :ok

Callback implementation for Mix.Task.run/1.