Mix v1.3.4 Mix.Tasks.Archive.Build

Builds an archive according to the specification of the Erlang Archive Format.

The archive will be created in the current directory (which is expected to be the project root), unless an argument -o is provided with the file name.

Archives are meant to contain small projects, usually installed locally. By default, this command archives the current project but the -i and -o options can be used to archive any directory. For example, mix archive.build with no options translates to:

mix archive.build -i _build/ENV/lib/APP -o APP-VERSION.ez

Command line options

-o - specify output file name. If there is a mix.exs, defaults to “APP-VERSION.ez”.

-i - specify the input directory to archive. If there is a mix.exs, defaults to the current application build.

--no-compile - skip compilation. Only applies when mix.exs is available.

Functions

run(args)

run(OptionParser.argv) :: :ok

Callback implementation for Mix.Task.run/1.