Mix v1.0.5 Mix.Tasks.Compile.Yecc

Compile Yecc source files.

When this task runs, it will check the modification time of every file, and if it has changed, the file will be compiled. Files will be compiled in the same source directory with a .erl extension. You can force compilation regardless of modification times by passing the --force option.

Command line options

--force - forces compilation regardless of modification times

Configuration

:erlc_paths - directories to find source files. Defaults to ["src"].

:yecc_options - compilation options that apply to Yecc’s compiler. There are many other available options here: http://www.erlang.org/doc/man/yecc.html#file-1.

Functions

clean()

Cleans up compilation artifacts.

manifests()

Returns Yecc manifests.

run(args)

Specs

run(OptionParser.argv) :: :ok | :noop

Runs this task.