Registry v0.2.0 Registry

A local, decentralized and scalable key-value process storage.

It allows developers to lookup one or more process with a given key. If the registry has :unique keys, a key points to 0 or 1 processes. If the registry allows :duplicate keys, a single key may point to 0, 1 or many processes. In both cases, a process may have multiple keys.

Each entry in the registry is associated to the process that has registered the key. If the process crashes, the keys associated to that process are automatically removed.

The registry can be used for name lookups, using the :via option, for storing properties, for custom dispatching rules as well as a pubsub implementation. We explore those scenarios below.

The registry may also be transparently partitioned, which provides more scalable behaviour for running registries on highly concurrent environments with thousands or millions of entries.

Using in :via

Once the registry is started with a given name, it can be used to register and access named processes using the {:via, Registry, {registry, key}} tuple:

iex> {:ok, _} = Registry.start_link(:unique, Registry.ViaTest)
iex> name = {:via, Registry, {Registry.ViaTest, "agent"}}
iex> {:ok, _} = Agent.start_link(fn -> 0 end, name: name)
iex> Agent.get(name, & &1)
0
iex> Agent.update(name, & &1 + 1)
iex> Agent.get(name, & &1)
1

Typically the registry is started as part of a supervision tree though:

supervisor(Registry, [:unique, Registry.ViaTest])

Only registry with unique keys can be used in :via. If the name is already taken, the start_link function, such as Agent.start_link/2 above, will return {:error, {:already_started, current_pid}}.

Using as a dispatcher

Registry has a dispatch mechanism that allows developers to implement custom dispatch logic triggered from the caller. For example:

iex> {:ok, _} = Registry.start_link(:duplicate, Registry.DispatcherTest)
iex> {:ok, _} = Registry.register(Registry.DispatcherTest, "hello", {IO, :inspect})
iex> Registry.dispatch(Registry.DispatcherTest, "hello", fn entries ->
...>   for {pid, {module, function}} <- entries, do: apply(module, function, [pid])
...> end)
# Prints #PID<...>
:ok

By calling register/3 different processes can register under a given key and associate a specific {module, function} tuple under that key. Later on, an entity interested in dispatching events, may call dispatch/3, which will receive all entries under a given key, allowing the each {module, function} to be invoked.

Keep in mind dispatching happens in the process that calls dispatch/3, registered processes are not involved in dispatching unless they are explicitly sent messages to. That’s the example we will see next.

Using as a PubSub

Registries can also be used to implement a local, non-distributed, scalable PubSub by relying on the dispatch/3 function, similar to the previous section, except we will send messages to each associated process, instead of invoking a given module-function.

In this example, we will also set the number of partitions to the number of schedulers online, which will make the registry more performant on highly concurrent environments and allowing dispatching to happen in parallel:

iex> {:ok, _} = Registry.start_link(:duplicate, Registry.PubSubTest,
...>                                partitions: System.schedulers_online)
iex> {:ok, _} = Registry.register(Registry.PubSubTest, "hello", [])
iex> Registry.dispatch(Registry.PubSubTest, "hello", fn entries ->
...>   for {pid, _} <- entries, do: send(pid, {:broadcast, "world"})
...> end)
:ok

The example above broadcasted the message {:broadcast, "world"} to all processes registered under the topic “hello”.

The third argument given to register/3 is a value associated to the current process. While in the previous section we used it when dispatching, in this particular example we are not interested on it, so we have set it to an empty list. You could store a more meaningful value if necessary.

Registrations

Looking up, dispatching and registering are efficient and immediate at the cost of delayed unsubscription. For example, if a process crashes, its keys are automatically removed from the registry but the change may not propagate immediately. This means certain operations may return process that are already dead. When such may happen, it will be explicitly written in the function documentation.

However, keep in mind those cases are typically not an issue. After all, a process referenced by a pid may crash at any time, including between getting the value from the registry and sending it a message. Many parts of the standard library are designed to cope with that, such as Process.monitor/1 which will deliver the DOWN message immediately if the monitored process is already dead and Kernel.send/2 which acts as a no-op for dead processes.

Types

key()

key :: term

kind()

kind :: :unique | :duplicate

registry()

registry :: atom

value()

value :: term

Functions

dispatch(registry, key, mfa_or_fun)

dispatch(registry, key, (entries :: [{pid, value}] -> term)) :: :ok

Invokes the callback with all entries under key in each partition for the given registry.

The list of entries is a non-empty list of two-element tuples where the first element is the pid and the second element is the value associated to the pid. If there are no entries for the given key, the callback is never invoked.

If the registry is partitioned, the callback will be invoked for every partition that has matching entries concurrently.

Keep in mind the dispatch/3 function may return entries that have died but have not yet been removed from the table. If this can be an issue, consider explicitly checking if the process is alive in the entries list. Remember there are no guarantees the process will remain alive, after all the process may also crash right after the Process.alive?/1 check.

See the module documentation for examples of using the dispatch/3 function for building custom dispatching or a pubsub system.

info(registry, key)

info(registry, info_key :: atom) ::
  {:ok, info_value :: term} |
  :error

Reads registry metadata given on start_link/3.

Examples

iex> Registry.start_link(:unique, Registry.InfoDocTest, custom_key: "custom_value")
iex> Registry.info(Registry.InfoDocTest, :custom_key)
{:ok, "custom_value"}
iex> Registry.info(Registry.InfoDocTest, :unknown_key)
:error

keys(registry, pid)

keys(registry, pid) :: [key]

Returns the known keys for the given pid in registry in no particular order.

If the registry is unique, the keys are unique. Otherwise they may contain duplicates if the process was registered under the same key multiple times. The list will be empty if the process is dead or it has no keys in this registry.

Examples

Registering under a unique registry does not allow multiple entries:

iex> Registry.start_link(:unique, Registry.UniqueKeysTest)
iex> Registry.keys(Registry.UniqueKeysTest, self())
[]
iex> {:ok, _} = Registry.register(Registry.UniqueKeysTest, "hello", :world)
iex> Registry.register(Registry.UniqueKeysTest, "hello", :later)
{:error, {:already_registered, self()}}
iex> Registry.keys(Registry.UniqueKeysTest, self())
["hello"]

Such is possible for duplicate registries though:

iex> Registry.start_link(:duplicate, Registry.DuplicateKeysTest)
iex> Registry.keys(Registry.DuplicateKeysTest, self())
[]
iex> {:ok, _} = Registry.register(Registry.DuplicateKeysTest, "hello", :world)
iex> {:ok, _} = Registry.register(Registry.DuplicateKeysTest, "hello", :world)
iex> Registry.keys(Registry.DuplicateKeysTest, self())
["hello", "hello"]

lookup(registry, key)

lookup(registry, key) :: [{pid, value}]

Finds the {pid, value} pair for the given key in registry in no particular order.

The pid-value pair is only returned if the pid is alive. An empty list will be returned if there is no alive pid registered under this key.

For unique registries, a single partition lookup is necessary. For duplicate registries, all partitions must be looked up.

Examples

In the example below we register the current process and look it up both from itself and other processes:

iex> Registry.start_link(:unique, Registry.UniqueLookupTest)
iex> Registry.lookup(Registry.UniqueLookupTest, "hello")
[]
iex> {:ok, _} = Registry.register(Registry.UniqueLookupTest, "hello", :world)
iex> Registry.lookup(Registry.UniqueLookupTest, "hello")
[{self(), :world}]
iex> Task.async(fn -> Registry.lookup(Registry.UniqueLookupTest, "hello") end) |> Task.await
[{self(), :world}]

The same applies to duplicate registries:

iex> Registry.start_link(:duplicate, Registry.DuplicateLookupTest)
iex> Registry.lookup(Registry.DuplicateLookupTest, "hello")
[]
iex> {:ok, _} = Registry.register(Registry.DuplicateLookupTest, "hello", :world)
iex> Registry.lookup(Registry.DuplicateLookupTest, "hello")
[{self(), :world}]
iex> {:ok, _} = Registry.register(Registry.DuplicateLookupTest, "hello", :another)
iex> Enum.sort(Registry.lookup(Registry.DuplicateLookupTest, "hello"))
[{self(), :another}, {self(), :world}]

put_info(registry, key, value)

put_info(registry, info_key :: atom, info_value :: term) :: :ok

Stores registry metadata.

Examples

iex> Registry.start_link(:unique, Registry.InfoDocTest, custom_key: "custom_value")
iex> Registry.info(Registry.InfoDocTest, :custom_key)
{:ok, "custom_value"}
iex> Registry.put_info(Registry.InfoDocTest, :custom_key, "new_value")
:ok
iex> Registry.info(Registry.InfoDocTest, :custom_key)
{:ok, "new_value"}

register(registry, key, value)

register(registry, key, value) ::
  {:ok, pid} |
  {:error, {:already_registered, pid}}

Registers the current process under the given key in registry.

A value to be associated with this registration must also be given. This value will be retrieved whenever dispatching or doing a key lookup.

This function returns {:ok, owner} or {:error, reason}. The owner is the pid of the registry partition responsible for the pid. The owner is automatically linked to the caller.

If the registry has unique keys, it will return {:ok, owner} unless the key is already associated to a pid, in which case it returns {:error, {:already_registered, pid}}.

If the registry has duplicate keys, multiple registrations from the current process under the same key are allowed.

Examples

Registering under a unique registry does not allow multiple entries:

iex> Registry.start_link(:unique, Registry.UniqueRegisterTest)
iex> {:ok, _} = Registry.register(Registry.UniqueRegisterTest, "hello", :world)
iex> Registry.register(Registry.UniqueRegisterTest, "hello", :later)
{:error, {:already_registered, self()}}
iex> Registry.keys(Registry.UniqueRegisterTest, self())
["hello"]

Such is possible for duplicate registries though:

iex> Registry.start_link(:duplicate, Registry.DuplicateRegisterTest)
iex> {:ok, _} = Registry.register(Registry.DuplicateRegisterTest, "hello", :world)
iex> {:ok, _} = Registry.register(Registry.DuplicateRegisterTest, "hello", :world)
iex> Registry.keys(Registry.DuplicateRegisterTest, self())
["hello", "hello"]

start_link(kind, registry, options \\ [])

start_link(kind, registry, Keyword.t) ::
  {:ok, pid} |
  {:error, term}

Starts the registry as a supervisor process.

Manually it can be started as:

Registry.start_link(:unique, MyApp.Registry)

In your supervisor tree, you would write:

supervisor(Registry, [:unique, MyApp.Registry])

For intensive workloads, the registry may also be partitioned. If partioning is required a good default is to set the number of partitions to the number of schedulers available:

Registry.start_link(:unique, MyApp.Registry, partitions: System.schedulers_online())

or:

supervisor(Registry, [:unique, MyApp.Registry, [partitions: System.schedulers_online()]])

The registry uses one ETS table for metadata plus 2 ETS tables per partition.

Options

options is a keyword list with metadata that is stored in the registry and may be looked up using the info/2 function. Only the following keys have meaning for the registry:

:partitions - the number of partitions in the registry. Defaults to 1.

All other keys are stored as is.

unregister(registry, key)

unregister(registry, key) :: :ok

Unregisters all entries for the given key associated to the current process in registry.

Always returns :ok and automatically unlinks the current process from the owner if there are no more keys associated to the current process.

Examples

It unregister all entries for key for unique registries:

iex> Registry.start_link(:unique, Registry.UniqueUnregisterTest)
iex> Registry.register(Registry.UniqueUnregisterTest, "hello", :world)
iex> Registry.keys(Registry.UniqueUnregisterTest, self())
["hello"]
iex> Registry.unregister(Registry.UniqueUnregisterTest, "hello")
:ok
iex> Registry.keys(Registry.UniqueUnregisterTest, self())
[]

As well as duplicate registries:

iex> Registry.start_link(:duplicate, Registry.DuplicateUnregisterTest)
iex> Registry.register(Registry.DuplicateUnregisterTest, "hello", :world)
iex> Registry.register(Registry.DuplicateUnregisterTest, "hello", :world)
iex> Registry.keys(Registry.DuplicateUnregisterTest, self())
["hello", "hello"]
iex> Registry.unregister(Registry.DuplicateUnregisterTest, "hello")
:ok
iex> Registry.keys(Registry.DuplicateUnregisterTest, self())
[]

update(registry, key, callback)

update(registry, key, (value -> value)) ::
  {new_value :: term, old_value :: term} |
  :error

Updates the value for key for the current process in the unique registry.

Returns the {old_value, new_value} or :error if there is no such key assigned to the current process.

If a non-unique registry is given, an error is raised.

Examples

iex> Registry.start_link(:unique, Registry.UpdateTest)
iex> {:ok, _} = Registry.register(Registry.UpdateTest, "hello", 1)
iex> Registry.lookup(Registry.UpdateTest, "hello")
[{self(), 1}]
iex> Registry.update(Registry.UpdateTest, "hello", & &1 + 1)
{2, 1}
iex> Registry.lookup(Registry.UpdateTest, "hello")
[{self(), 2}]