Elixir v1.1.1 MapSet

A set store.

The MapSet is represented internally as a struct, therefore %MapSet{} can be used whenever there is a need to match on any MapSet. Note though the struct fields are private and must not be accessed directly. Instead, use the functions on this or in the Set module.

The MapSet is implemented using Map data type. For more information about the functions and their APIs, please consult the Set module.

Functions

delete(set, term)

Callback implementation for Set.delete/2.

difference(set1, set2)

Callback implementation for Set.difference/2.

disjoint?(set1, set2)

Callback implementation for Set.disjoint?/2.

equal?(map_set1, map_set2)

Callback implementation for Set.equal?/2.

intersection(set1, set2)

Callback implementation for Set.intersection/2.

member?(map_set, value)

Callback implementation for Set.member?/2.

new()

Callback implementation for Set.new/0.

put(set, value)

Callback implementation for Set.put/2.

size(map_set)

Callback implementation for Set.size/1.

subset?(set1, set2)

Callback implementation for Set.subset?/2.

to_list(map_set)

Callback implementation for Set.to_list/1.

union(map_set1, map_set2)

Callback implementation for Set.union/2.