Mix v1.3.4 Mix.Tasks.Archive.Install

Installs an archive locally.

If no argument is supplied but there is an archive in the project’s root directory (created with mix archive.build), then the archive will be installed locally. For example:

mix do archive.build, archive.install

The argument can be an archive located at some URL:

mix archive.install http://example.com/foo.ez

After installation, the tasks in the archive are available locally:

mix some_task

Command line options

--sha512 - checks the archive matches the given SHA-512 checksum

--force - forces installation without a shell prompt; primarily intended for automation in build systems like make

Functions

check_path_or_url(path_or_url)

Callback implementation for Mix.Local.Installer.check_path_or_url/1.

find_previous_versions(src, dst)

Callback implementation for Mix.Local.Installer.find_previous_versions/2.

install(ez_dst, contents, previous)

Callback implementation for Mix.Local.Installer.install/3.

run(argv)

run(OptionParser.argv) :: boolean

Callback implementation for Mix.Task.run/1.