Elixir v1.1.1 Port

Functions related to Erlang ports.

Functions

call(port, operation, data)

See :erlang.port_call/3

Inlined by the compiler.

close(port)

See :erlang.port_close/1

Inlined by the compiler.

command(port, data, options \\ [])

See :erlang.port_command/2

Inlined by the compiler.

connect(port, pid)

See :erlang.port_connect/2

Inlined by the compiler.

control(port, operation, data)

See :erlang.port_control/3

Inlined by the compiler.

info(port)

Returns information about the port or nil if the port is closed.

See :erlang.port_info/1

info(port, spec)

Specs

info(port, atom) :: {atom, term} | nil

Returns information about the port or nil if the port is closed.

See :erlang.port_info/2

list()

See :erlang.ports/0

Inlined by the compiler.

open(name, settings)

See :erlang.open_port/2

Inlined by the compiler.